[Task] Adds Design To Layout And Specs For The Current Setup
This commit adds css to design basic layouts and specs for tweet model
and controller. Also encountered an issue with rspec-rails version being
non-compatible with Rails 6, hence as a solution have updated the gem
Rspec to version 4 which is currently in development, more details for
this can be found in the Gemfile.